+++ example/src/@types/models/user.d.ts
@@ -1,0 +1,6 @@
+interface IUser {
+  _id: import('mongoose').Types.ObjectId
+  name: string
+  lastName: string
+  updatedAt: Date
+}