+++ example/src/controllers/utils/messages/index.ts
@@ -1,7 +1,0 @@
-import { EFU, MFU } from './user'
-
-enum GenericErrors {
-  INTERNAL_SERVER_ERROR = 'Something went wrong'
-}
-
-export { EFU, MFU, GenericErrors as GE }