+++ example/src/database/mongo/models/user.ts
@@ -1,0 +1,34 @@
+import { model, Schema } from 'mongoose'
+
+const User = new Schema<IUser>(
+  {
+    lastName: {
+      required: true,
+      type: String
+    },
+    name: {
+      required: true,
+      type: String
+    }
+  },
+  {
+    timestamps: {
+      createdAt: false,
+      updatedAt: true
+    },
+    toJSON: {
+      transform(_, ret) {
+        ret.id = ret._id
+        delete ret._id
+        delete ret.__v
+        delete ret.updatedAt
+      },
+      versionKey: false,
+      virtuals: true
+    }
+  }
+)
+
+const UserModel = model<IUser>('users', User)
+
+export { UserModel }