+++ example/src/dto-interfaces/index.ts
@@ -1,3 +1,0 @@
-import { DtoUser } from './user'
-
-export { DtoUser }