+++ example/src/models/index.ts
@@ -1,3 +1,0 @@
-import { IUser, UserModel } from './user'
-
-export { IUser, UserModel }