+++ example/src/network/router.ts
@@ -2,7 +2,8 @@ import { Application, Response, Request, Router, NextFunction } from 'express'
 import swaggerUi from 'swagger-ui-express'
 import httpErrors from 'http-errors'
 
-import { Home, User } from 'routes'
+import { response } from './response'
+import { Home, User } from './routes'
 import { docs } from 'utils'
 
 const routers = [User]