+++ example/src/routes/index.ts
@@ -1,4 +1,0 @@
-import { Home } from './home'
-import { User } from './user'
-
-export { Home, User }